stain-classification-api/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.0
python-multipart==0.0.9
onnxruntime-gpu==1.19.2
numpy==1.26.4
Pillow==10.3.0
python-dotenv==1.0.1
